supabase/requirements.txt
# Requirements for Agent Swarm Knowledge System

# HTTP client
httpx>=0.25.0

# OpenAI API
openai>=1.0.0

# Async support
asyncio>=3.4.3

# Type hints
typing-extensions>=4.0.0

# Date/time parsing
python-dateutil>=2.8.0

# XML parsing (for RSS)
lxml>=4.9.0

# Feed parsing
feedparser>=6.0.0

# HTML parsing (for web scraping)
beautifulsoup4>=4.12.0

# Environment variables
python-dotenv>=1.0.0

# Retry logic with exponential backoff
tenacity>=8.2.0

# Circuit breaker pattern for API resilience
pybreaker>=1.0.0

# Configuration management
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-httpx>=0.27.0

# Code quality
flake8>=6.0.0
black>=23.0.0
mypy>=1.5.0
